pyproject.toml
[tool.poetry]
name = "openfabric-test"
version = "1.1"
description = "cognitive-assistant"
authors = ["Andrei"]

[tool.poetry.dependencies]
python = "^3.8"
openfabric-pysdk = "0.2.7"
transformers = "^4.39.2"
torchvision = "^0.17.2"
tokenizers = "^0.15.2"
bitsandbytes = "^0.43.0"
protobuf = "^5.26.1"
accelerator = "^2024.3.8.dev1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
